Fixes:    Fixed duplicate check when adding files using the resource editor
Task:     238308
RevBy:    eike
AutoTest: -
Details:  Instead of the text and prefix of the file to add, pointers were compared. Thus the function always returned false and duplicates were not recognized. Replaced struct FileList with a class FileList, which is inherited from QList and added the function containsList which checks the needed values.