Fixes:    duplicate check when adding resources using the resource editor fixed
Task:     -
RevBy:    -
AutoTest: -
Details:  If the user wants to add a resource which is not located in the directory/a subdirectory of the qrc-file, he is asked what should be done. If he chose to copy the file to another directory, the check, if this directory is valid ((sub-)directory of the qrc-file) was missing. Added it.