Fixes:    Workaround a bug in kde 4.2.0.
Task:     Reported on irc.
Details:  We did regiser all plugins but didn't initialize them. Leading
to the kde plugins installing a kde file dialog, but not initializing a
KComponentData, which leads to a crash when trying to open a file
dialog. So we aren't lazy anymore with the designer plugins.
And initialize them right away. We should reenable lazy initializing
once kde 4.2.1 has been released. (And probably has replaced kde 4.2.0
on most systems.)