Stop fighting over which mouse cursor to show on text editor
The BaseTextEditor was unconditionally setting the mouse cursor on each
mouse move event, after which the CPPEditor would set it as well when
the mouse was above a link. This caused some mouse cursor flickering, so
now the cursor is only set when it's supposed to change.

Also fixed an issue where the link wasn't removed when leaving the text
editor with the mouse while Ctrl was pressed.