re-use pre-existing builds
run "cmake . [...]" if a build already exists in the build dir. this
saves the user from reproducing all the arguments he may need.

on the downside, this will go utterly awry if the source dir of the
existing build is a different one than known to creator. this should be
a rather rare corner case, so ignore it for now. the nice solution would
be a checkbox (on by default) to request this new behavior.