Add a next/prev shortcut (F6 for now) for outputpanes.
This one has some magic inside to always use the current outputpane.
This solves the issue that I don't want Search and Build Issues list to
have its own set of next/prev shortcuts. todo is a better definition of
next for the build issues and/or a way to restrict it only to errors.