beautify installation of pluginspecs
use the data copying custom compiler approach here, too.
this leaves more work to qmake, which is good.
on the upside, pluginspecs will be installed when they change without
having force a re-link of the plugin itself.
on the downside, installation of a pluginspec will force a re-link of
the plugin. oh, well - you can't have it all (with qmake) ...