Introduced settings database based on sqlite
Potential replacement for QSettings. This database will allow us to save
settings incrementially, removing most of the performance issues we're
having when our few megabytes of settings are being saved.

Not done in the form of a QSettings backend because it doesn't seem the
backends can do an incremental update of the settings.

Currently used by the QuickOpenPlugin for testing.