Fixed indenting of else line when typing "} else {"
The line would only get a reduced indent when the user actually typed
the }, but not when typing the {, even though the line should still be
unindented because of the } at the start.

Previously this code worked fine because lines didn't used to get
automatically reindented for anything else than the first character.