Make FakeVim use the default find box instead of mini-buffer
Creator already has a find box that works fairly well, so there
is no need for a mini-buffer mode for searching.  This change makes
FakeVim activate the standard find box when the user hits '/' or '?'.

Note: right now both '/' and '?' do forward searching by default
because the Creator find box lacks a "find backwards by default" mode.
This can be added later to the find box if considered useful.