Fixes identifing the right WD for .pro chaning debug_and_release
CONFIG+=debug_and_release not working on linux and
CONFIG-=debug_and_release not working on windows.
The logic should be now correct. That is the profile reader reads the
values that are set for the qt build, gets qmake arguments which we
might pass on the command line and reads the .pro files, and in the end
we are interested what CONFIG contains. Patch from Michael Karcher

Task-number: 254925