Add a copy full path to clipboard to open edtiors combobox
As requested on irc, and another common IDE has it at the same place. I
wonder whether we should have more options in that context menu. I
mainly put this in, because we don't have a good way to copy that full
path and I don't see a particular nice way to add it. Though I dislike
that it is only available in the context menu.