Always show the wizard even if we already have a recent enough cbp file.
That case is a corner case, it can happen if the user has run the codeblocks generator on the command line or if the user deleted the .user file. We mostly ask to get the user a chance to check that we set the correct generator. Or let him select the correct one if there are multiple msvc versions.