Only call setVisible on this item after it has been put in a layout
If you call setVisible(true) on the item before it is placed in a
hierarchy, it will show as a normal top-level window. On some platforms
this isn't really an issue, but with Carbon, the show happens
immediately, so you get this flash of a window which is really
disturbing. I'm surprised it's not noticed on other platforms.
Personally, I'm not sure why we call this at all, but I'm leaving it for
the moment.