Made trailing slash trigger the next include completion
The completion for the directory wasn't triggered because one of the
items matched the typed string, so the list of completion wasn't empty
and no new trigger was allowed.

By not matching items when a trailing slash is typed, the list is
allowed to trigger again from the new location.

Reviewed-by: Daniel Molkentin <daniel.molkentin@nokia.com>