Fix not trying hard enough to find the correct file on build errors.
If there are multiple files with the same name, then we try harder to
find the correct one. That is for jom and nmake we have a relative path
so we try to find a file which matches the relative path completly.
Instead of just comparing the last part.
(cherry picked from commit 156db185922238a3a214012cf406efba2b6c8b2c)