debugger: small refactoring of the access to the gdb process in the gdbengine
That's the first step to make the gdb/trk Adapter work with the otherwise
unmodified gdb engine. GdbProcessBase is inherited by GdbProcess and the
gdb/trk Adapter. In the GdbProcess case it's just a wrapper around a QProcess
running gdb, in the Adapter case it's the interface to the gdb process in the
whole rfomm/gdb/gdbserver combo.