debugger: replace hack to select trk adapter and pass parameters by a new one
Now .pro files can be used as project description as usual. To make the
debugger pick up the "on device" part a magic commandline is needed:
"@sym@ c:\sys\bin\your.exe  /path/to/your.sym"

If the project is not compilable, remove the Make and QMake buildsteps
and add a new Custom step that's only calling /bin/true.