Fixed transient compile failure with highly parallel jom.
These .pri files which add to INCLUDEPATH and have FORMS should also add
to DEPENDPATH.

Adding to INCLUDEPATH does not add to DEPENDPATH by default with qmake.
That means qmake may not find headers in those paths and won't
consider them dependencies when compiling the source files which include
them.

This doesn't matter when all the headers in the new INCLUDEPATH are
always present, but when some of them are generated (e.g. from FORMS),
the dependencies are important and omitting them means a race condition
in parallel builds.

Fixes errors of the form:
  ..\..\shared\help\topicchooser.h(33) : fatal error C1083: Cannot open
    include file: 'ui_topicchooser.h': No such file or directory