Debugger: Do configuration error checking early on.
Add a configuration checking method to the Debugger manager,
depending on toolchain, wire it to the engines.
Check that in the debugger run controls.

Add a convenience method to ICore that shows a warning
message with a "Settings" button, pointing the user
to a configuration error on a settings page.

Remove leftovers of the dumper parser.
Acked-by: con <qtc-committer@nokia.com>
(cherry picked from commit 22ab8d5662590a2150d6c4ba6472c201f5914181)

Conflicts:

	src/plugins/debugger/gdb/gdbengine.cpp