debugger: fix issue where child items oeverwrite parent items in the watch model.
The real problems is somewhere in the watch model. This patch does
only prevent the wrong (and unneeded) questions to be asked in the first
place. It does not fix the wrong handling of the answer in the watch model.

Reviewed-by: Oswald Buddenhagen
(cherry picked from commit 4ceabf26d4c0ba0bf023c7f7605bca639f338c48)