Fixed search results handling of font height
This patch makes sure that the rows that show the file and the rows that
show the results are the same height, to satisfy the uniform row heights
setting.

Also, a layout changed signal is now emitted when the font is changed.
This makes active search results update correctly when the text editor
font is changed.

Reviewed-by: con