Make context help work also when switching user interface language.
Also avoids errors because of users messing with the "Unfiltered"
filter, which is used for e.g. context help.
The strategy now is to remove the "Unfiltered" filter from previous runs
(possibly in a different language), and re-register a fresh filter with
the name in the current language, and guaranteed to not filter out
anything.

Reviewed-by: kh1
(cherry picked from commit ad4c227ca45912b27383236d8b73b790e97acf8e)