Fixed the sorting of search results
The search results window assumed the files were searched in sorted
order, which is not always the case especially since "find usages" can
now operate in several threads.

This change makes sure that the list of files in the search results is
always sorted, using a binary search to find the right index to insert
new files.

Reviewed-by: con