Be smarter about CamelCase when doing case-insensitive completion
Even when an upper case character can also match a lower case one, it
should only be treated as a CamelCase match when the matched character
is also upper case. This avoids some false positives.

Also, the part that allows a word in a CamelCase identifier to be
truncated should never be matched case-insensitively.