Fixed disabled action shortcuts
The patch introduces a dummy action which eats the respective
shortcuts when the real action is disabled.
This works around an issue on X11, where some Ctrl-shortcuts do
insert printable text, for example Ctrl+. for the next bookmark.
Without the patch, Ctrl+. would insert an unexpected '.' into
the text when there are no bookmarks.

Reviewed-by: con