Fixed crash when leaving session with invalid bookmarks or breakpoints
When the bookmark could not be added to the editor due to being on a
non-existing line, it would not be cleaned up properly when the editor
was closed, resulting in a crash when it later tried to remove itself
from the no longer existing editor.

In addition to fixing the crash, bookmarks that are not on valid lines
are now automatically removed when you try to navigate to them.

Task-number: QTCREATORBUG-545
Reviewed-by: mae
(cherry picked from commit 2b46f828b4ae2f41c4efc9fde7e8b7c5e271b4de)