Let the DesignMode recover when the qml contains an error on first load
If the qml contains an error on frist call of
DesignDocumentController::loadMaster, we did not wire up the components
to the model. If the user then fixed the qml, the Design mode widgets
were 'enabled', but didn't contain any meaningful input.

Fixed by always wiring up the components. Unfortunately this shows the
'default' model with just a Rectangle.