Move BuildStep::init() to where they are added to the buildqueue
Actually that's more correct. E.g. imagine building a project, and while
the qmake step runs changing the working directory. The make step would
be affected by that. With the change both the qmake and make step get
their working directory on beeing added to the queue and so don't care
if the setting is changed afterwards.