Opening binary files leaves empty entries in open editors window
Leading to a crash when selecting the empty entry.
The patch fixes a big memory leak with bin editors (neither the widget
nor the IFile instance was deleted, the latter leading to the empty
entry in the open editors window), and also puts a saveguard in the open
editors window, so that empty entries are not shown any more.

Task-number: QTCREATORBUG-571
Reviewed-by: Thorbjørn