Fixed performance issue with scrolling when using X11 graphics system
Depending on the XRender driver (probably), drawing a short straight
line can take quite a while when this line is transparent. On my
laptop's onboard ATI card, this meant that scrolling was unusably slow
when this transparent line was being drawn.

Fixed by avoiding the issue, since there is no actual need to use
transparency here. We can calculate the resulting color in advance
instead.