Fix assert in WriteLogger because of misbehaving Navigator
The Navigator "called back" to the model in case e.g. the id
of an item changed. This was catched by the WriteLogger, which
lead to an assert.

Fix this by tracking whether we are changing the navigator items
due to a model change, and not reacting to itemChanged signal
in this case.