Fixed cursor position in file when switching editors
When closing editor splits while the same document was visible in
two editors, it could happen that the state of the closed editor
survived (as opposed to sate of the editor which was still visible).
This had the negative effect, that when visiting the file again,
the cursor would jump to the obsolete position within the file.