Disable adding of files to a .qmlproject
The old code still assumed the old, plain text format, and just added
the file path to the end of the file. This broke the new .qmlproject
file format.

Ideally QmlProject should be clever enough to check whether the file
will be automatically picked up in the monitored directories, and if
not added explicitly e.g. to QmlFiles { paths: ... }. Anyhow, this
quick fix just shows every time an error message instead of destroying
the .qmlproject file.