QmlDesigner: Fix artifact when moving items around
Don't use the dynamic property "HasFormEditorItem" any more when
checking whether childs have to be drawn directly by it's parent.

The problem was actually introduced with the new dynamic metatype
object (NodeInstanceMetaObject). At the time the dynamic property
"HasFormEditorItem" is set in
GraphicsObjectNodeInstance::createGraphicsObject(), the new metatype
object hasn't been created & attached to the object yet. The dynamic
property was effectively lost when the metatype object changed, leading
to multiple drawings of the child item.

This fix replaces the "HasFormEditorItem" property by directly asking
the NodeInstanceView whether an item draws itself or not.

Reviewed-by: Thomas Hartmann