Fix button toggle of the miniproject selector on Windows
The problem is that on windows, mouse events are replayed
if you click outside a popup to close it. If you hit the
project selector button to toggle it, this will cause
the popup to reopen on mouse release. Which we do not want.
Hence we set the DontReplay flag on mouseClick.

A similar workaround can be found in QComboBox.