Force custom colors within usable range
We previously allowed fully white and yellow colors as the
base color for our interface. This does not work with our
icons or other interface elements. Instead I now constrain
the selectable colors within a certain range of saturation
and brightness. This still leaves the user in control but
will prevent the common case where the user selects an
overly saturated color from the color picker and ends up
with an unusable or extremely ugly panel.