Enable suppressing syntax and setting a foreground color on disabled blocks
If the foreground color for the disabled text is different than the normal
text foreground, that's an indication that the user wants the disabled
block to be colored differently. If so, disable the syntax highlighting
and set uniform colors for the block.

If the foreground colors are identical apply full syntax highlighting in
the disabled block too. Some people want that.

Reviewed-by: Thorbjørn Lindeijer