Fix crash for Custom Executable RunConfiguation with qml projects
Custom Executable RunConfiguation assumed that there is always a active
buildconfiguration which is simply not true for qml projects.

Also fix a small bug that the executable was not searched in the full
environment but only in the build environment.

Reviewed-By: Tobias Hunger
(cherry picked from commit e35a1e612a8817b3393967d7aea9e86564c210cc)