Fixed crash when trying to activate a no longer existing editor
This could happen when using Ctrl+Tab to switch editors, and while
holding Ctrl you close an editor with Ctrl+W, and then you try to select
it.

Since the crash was due to an assertion, we can just avoid it by doing
nothing when no editor for the given file is found.

Reviewed-by: mae
Task-number: QTCREATOR-190