QmlDesigner: Fix crash when changing the transform origin of an item
Fixes crash when changing the transform origin of an item in the
Property Editor. The fundamental problem is that the metatype system
doesn't know about the enum. We used to return then an invalid QVariant
when trying to convert the string to a QVariant. Now we instead return
a valid QVariant, but with wrong type (QString).

Task-number: BAUHAUS-522
Reviewed-by: Erik Verbruggen