Allow loading of projects even with no valid Qt
This patch marks up invalid Qt versions as Desktop.

Make sure we add a the always existing invalid Qt version in the path
available in the targetselectionpage if no valid Qt versions could be found.

This way a project can be set up and loaded without a valid Qt version set up.
Trying to build this will fail of course. This is basically how we handled
this situation before we had targets, too.

Task-number: QTCREATORBUG-1066
Reviewed-by: dt