"Fix" overwrite cursor for fake vim mode
It now behaves less logical but closer to what vim does. Since
this is the main target group, the change is probably right.
In detail: Overwrite cursor no longer shows *what* will get overwritten,
but instead it simply indicates *that* the editor is in overwrite mode,
with a more or less fixed size (fixed when the font is truly monospace).