Avoid build progress' error widget flickering, and popping up as toplevel on Mac
When a build was started after a build had errors, the error widget
would be made visible before it had a parent, leading to it being shown
as a top level for a split second.
So, clear the tasks before creating the new error widget, and assume
that we start without errors in the error widget.