QmlDesigner: fixes crash in nodeinstances
There was a subtle bug in propertyReseting() that caused a gradient
property of a Rectangle not to be removed, if the gradient was deleted.
Deleting the gradient triggered a reparenting operation in the
nodeinstances where the resetProperty() of the oldParent was not
executed properly.

This caused a dangling pointer in QDeclarativeRectangle.

Reviewed-by: Marco Bubke