QuickDesigner: Reduce numbers of redraws / state switching in States Editor
nodeInstancePropertyChanged() is called for _every_ state switch ... so
the current assumption that it only reflects real changes does not hold.

Instead, use the orphaned sceneChanged() method to schedule redraws.
However, this still results in some overhead when switching to the base state
(sceneChanged() cannot decide whether it has to redraw all other states
or not).

Reviewed-by: Christiaan Janssen