HACK: ensure a cleaner sourcedir for symbian builds
The symbian makefile generator don't support shadow builds. Thus they
always pollute the source directory with build artefacts. Mixing
in-source and shadow builds is obviously broken.

We do as good as we can, setting OBJECTS_DIR, MOC_DIR, UI_DIR, RCC_DIR
to .obj, .moc, .ui and .rcc, which keeps the source dir clean of at least
some of the artefacts. This is as good as we can, but not really good.

Task-Nr: QTCREATORBUG-1293