Always reset the id for an object in node instances
Because we tested for the parent, the id of a root item never reset.
This was hidden because if you remove a root item all
child items was removed too. But if the object was not deleted by the
root item and there was a binding the root item id, there was a dangling
pointer.

Task-Number: BAUHAUS-662
Reviewed-by: kkoehne